feat: upgrade all project deps
This is a bigger commit which upgrades all project deps.
It also adds Wireit into the package.json for later as this
repo is planned to also house the code for a standalone
desktop app and using Wireit will make management easier.
This also adopts for SvelteKit 1.0 and other more breaking changes.

Signed-off-by: Marvin Drees <marvin.drees@9elements.com>