refactor: switch from carbon to skeleton
This is a giant rework of the base repo as it switches
the underlying framework out for Skeleton. The rationale here
is that carbon integrates not as good as skeleton and as the
repo is still rather unpopulated so doing this early on will
save us headaches later. This also removes the blocker for
full PnP zero-install support which will be changed in this
commit together with the changed UI framework.

Signed-off-by: Marvin Drees <marvin.drees@9elements.com>